Update command checks to make bot seems smarter

+++ commands/help.js
@@ -7,10 +7,13 @@ export const name = "help";
 export const execute = async (client, message, args) => {
   const DURATION = 60000;
   const data = await getAllCommands();
+  let detailedHelpCmd;
+
   const embedMsgs = [
-    new MessageEmbed().setColor("#2576A3")
+    new MessageEmbed()
+      .setColor("#2576A3")
       .setTitle("COMMANDS")
-      .setFooter(`Living in AWS EC2  \u2022  Page 1 / ${data.length}`, client.user.avatarURL())
+      .setFooter(`Living in AWS EC2  \u2022  Page 1 / ${data.length + 1}`, client.user.avatarURL())
       .setTimestamp()
       .setDescription(trimExtraSpace(`
         Use \`help <Command Name>\` for more information about that command.
@@ -27,36 +30,65 @@ export const execute = async (client, message, args) => {
   ];
 
   data.forEach((cmd, i) => {
+    if (args[1] && args[1] === cmd.command && args[1] !== "help") {
+      detailedHelpCmd = cmd;
+      return;
+    }
+
     embedMsgs.push(
-      new MessageEmbed().setColor("#2576A3")
+      new MessageEmbed()
+        .setColor("#2576A3")
         .setTitle(cmd.command.toUpperCase())
-        .setFooter(`Living in AWS EC2  \u2022  Page ${i + 2} / ${data.length}`, client.user.avatarURL())
+        .setFooter(`Living in AWS EC2  \u2022  Page ${i + 2} / ${data.length + 1}`, client.user.avatarURL())
         .setTimestamp()
         .setDescription(trimExtraSpace(`
           ${cmd.description}
 
-          **USAGE**
-          \u2022 ${cmd.usage.replace(/::/gm, "\n\u2022")}
+          **Usage**
+          \u2022 ${cmd.usage.replace(/::/gm, "\n\u2022 ")}
         `))
-    )
-  })
-
-  message.channel.send(embedMsgs[0]).then(async msg => {      
-    await msg.react("⬅️");
-    await msg.react("➡️");
-
-    let currentPage = 0;
-    const filter = (reaction, user) => (reaction.emoji.name === "⬅️" || reaction.emoji.name === "➡️") && !user.bot && user.id === message.author.id;
-    const collector = msg.createReactionCollector(filter, { time: DURATION, dispose: true });
-    
-    collector.on("collect", (reaction, user) => currentPage = updateEmbedPage(msg, embedMsgs, currentPage, reaction, user));
-    collector.on("remove", (reaction, user) => currentPage = updateEmbedPage(msg, embedMsgs, currentPage, reaction, user));
-
-    collector.on("end", () => {
-      msg.delete();
-      console.log(chalk.yellow("Embed help message deleted.\n"));
-    });
+    );
   });
+
+  if (!detailedHelpCmd) {
+    message.channel.send(embedMsgs[0]).then(async msg => {      
+      await msg.react("⬅️");
+      await msg.react("➡️");
+  
+      let currentPage = 0;
+      const filter = (reaction, user) => (reaction.emoji.name === "⬅️" || reaction.emoji.name === "➡️") && !user.bot && user.id === message.author.id;
+      const collector = msg.createReactionCollector(filter, { time: DURATION, dispose: true });
+      
+      collector.on("collect", (reaction, user) => currentPage = updateEmbedPage(msg, embedMsgs, currentPage, reaction, user));
+      collector.on("remove", (reaction, user) => currentPage = updateEmbedPage(msg, embedMsgs, currentPage, reaction, user));
+  
+      collector.on("end", () => {
+        try {
+          msg.delete();
+          console.log(chalk.yellow("Embed help message deleted.\n"));
+        }
+        catch (e) {
+          console.log(chalk.yellow("Embed message might have been delete already."));
+          console.log(chalk.yellow(`${e.name}: ${e.message}\n`));
+        }
+      });
+    });
+  }
+  else {
+    const embed = new MessageEmbed()
+      .setColor("#2576A3")
+      .setTitle(detailedHelpCmd.command.toUpperCase())
+      .setFooter(`Living in AWS EC2 ${data.length + 1}`, client.user.avatarURL())
+      .setTimestamp()
+      .setDescription(trimExtraSpace(`
+        ${detailedHelpCmd.description}
+
+        **Usage**
+        \u2022 ${detailedHelpCmd.usage.replace(/::/gm, "\n\u2022 ")}
+      `));
+
+    message.channel.send(embed);
+  }
 }
 
 const updateEmbedPage = (msg, embedMsgs, currentPage, reaction, user) => {
@@ -66,7 +98,7 @@ const updateEmbedPage = (msg, embedMsgs, currentPage, reaction, user) => {
     page--;
     msg.edit(embedMsgs[page]);
   }
-  else if (reaction.emoji.name === "➡️" && page < embedMsgs.length) {
+  else if (reaction.emoji.name === "➡️" && page < embedMsgs.length - 1) {
     page++;
     msg.edit(embedMsgs[page]);
   }
